refactor: split resource reader
