+++ campaign.py
@@ -9,7 +9,7 @@ import inspect
 
 import script.common as common
 import script.hud as hud
-import script.resources as resources
+import script.resources.reader as resources
 import script.screen_utils as screen_utils
 from script.config_utils import parse_scenario_info
