+++ campaigns/Ascent_of_Egypt/Egypt_1_Hunting.py
@@ -17,7 +17,7 @@ import os
 
 import script.common as common
 import script.hud as hud
-import script.resources as resources
+import script.resources.reader as resources
 import script.input_utils as input_utils
 from script.units import villager
 from script.buildings.town_center import train_villagers
