+++ script/units/villager.py
@@ -3,7 +3,7 @@ import time
 
 import script.common as common
 import script.hud as hud
-import script.resources as resources
+import script.resources.reader as resources
 import script.input_utils as input_utils
 
 logger = logging.getLogger(__name__)
